+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ### Unreleased
 
+### Maintenance
+
+* Removed `requirements.txt`; it was unnecessary and out of date
+* Replaced `requests` dependency with `httpx`
+
 
 ## [v1.10.0]